+++ test/010_math_permtoindex/permtoindex_tests.c
@@ -11,7 +11,7 @@ void run(void) {
 	n = atoll(str);
 	for (i = 0; i < n; i++) {
 		fgets(str, STRLENMAX, stdin);
-		a[i] = atoi(str);
+		a[i] = (uint8_t)atoi(str);
 	}
 
 	p = permtoindex(n, a);
