+++ test/013_math_digitstosumzero/digitstosumzero_tests.c
@@ -11,10 +11,10 @@ void run(void) {
 	fgets(str, STRLENMAX, stdin);
 	n = atoi(str);
 	fgets(str, STRLENMAX, stdin);
-	b = atoi(str);
+	b = (uint8_t)atoi(str);
 	for (i = 0; i < n; i++) {
 		fgets(str, STRLENMAX, stdin);
-		a[i] = atoi(str);
+		a[i] = (uint8_t)atoi(str);
 	}
 
 	p = digitstosumzero(n, a, b);
